[docker] refs #1433 Fix some scripts tests

+++ ci-scripts/integration-test-disable-gui.sh
@@ -11,10 +11,11 @@ while $(lsof -Pi :$PORT -sTCP:LISTEN -t >/dev/null) ; do
     PORT=$((PORT+1))
 done
 
+SKYCOIN_NODE="${SKYCOIN_NODE:-"127.0.0.1"}"
 COIN="${COIN:-skycoin}"
 RPC_PORT="$PORT"
-HOST="http://127.0.0.1:$PORT"
-RPC_ADDR="http://127.0.0.1:$RPC_PORT"
+HOST="http://$SKYCOIN_NODE:$PORT"
+RPC_ADDR="http://$SKYCOIN_NODE:$RPC_PORT"
 MODE="disable-gui"
 BINARY="${COIN}-integration-disable-gui.test"
 TEST=""
@@ -56,46 +57,52 @@ go test ./src/api/integration/...
 go test ./src/cli/integration/...
 
 DATA_DIR=$(mktemp -d -t ${COIN}-data-dir.XXXXXX)
-WALLET_DIR="${DATA_DIR}/wallets"
+WALLET_DIR="/wallet"
+
+if [ "$SKYCOIN_NODE" = "127.0.0.1" ]; then
+  WALLET_DIR="${DATA_DIR}/wallets"
+fi
 
 if [[ ! "$DATA_DIR" ]]; then
   echo "Could not create temp dir"
   exit 1
 fi
 
-# Compile the skycoin node
-# We can't use "go run" because that creates two processes which doesn't allow us to kill it at the end
-echo "compiling $COIN with coverage"
-go test -c -tags testrunmain -o "$BINARY" -coverpkg="${COVERPKG}/..." ./cmd/${COIN}/
+if [ "$SKYCOIN_NODE" = "127.0.0.1" ]; then
 
-mkdir -p coverage/
+  # Compile the skycoin node
+  # We can't use "go run" because that creates two processes which doesn't allow us to kill it at the end
+  echo "compiling $COIN with coverage"
+  go test -c -tags testrunmain -o "$BINARY" -coverpkg="${COVERPKG}/..." ./cmd/${COIN}/
 
-# Run skycoin node with pinned blockchain database
-echo "starting $COIN node in background with http listener on $HOST"
+  mkdir -p coverage/
 
-./"$BINARY" -disable-networking=true \
-            -web-interface-port=$PORT \
-            -download-peerlist=false \
-            -db-path=./src/api/integration/testdata/blockchain-180.db \
-            -db-read-only=true \
-            -launch-browser=false \
-            -data-dir="$DATA_DIR" \
-            -wallet-dir="$WALLET_DIR" \
-            -enable-all-api-sets=true \
-            -enable-api-sets=INSECURE_WALLET_SEED \
-            -enable-gui=false \
-            -test.run "^TestRunMain$" \
-            -test.coverprofile="${COVERAGEFILE}" \
-            &
+  # Run skycoin node with pinned blockchain database
+  echo "starting $COIN node in background with http listener on $HOST"
 
-SKYCOIN_PID=$!
+  ./"$BINARY" -disable-networking=true \
+              -web-interface-port=$PORT \
+              -download-peerlist=false \
+              -db-path=./src/api/integration/testdata/blockchain-180.db \
+              -db-read-only=true \
+              -launch-browser=false \
+              -data-dir="$DATA_DIR" \
+              -wallet-dir="$WALLET_DIR" \
+              -enable-all-api-sets=true \
+              -enable-api-sets=INSECURE_WALLET_SEED \
+              -enable-gui=false \
+              -test.run "^TestRunMain$" \
+              -test.coverprofile="${COVERAGEFILE}" \
+              &
 
-echo "$COIN node pid=$SKYCOIN_PID"
+  SKYCOIN_PID=$!
 
-echo "sleeping for startup"
-sleep 3
-echo "done sleeping"
+  echo "$COIN node pid=$SKYCOIN_PID"
 
+  echo "sleeping for startup"
+  sleep 3
+  echo "done sleeping"
+fi
 set +e
 
 if [[ -z $TEST || $TEST = "api" ]]; then
@@ -107,13 +114,16 @@ API_FAIL=$?
 
 fi
 
-echo "shutting down $COIN node"
+if [ "$SKYCOIN_NODE" = "127.0.0.1" ]; then
 
-# Shutdown skycoin node
-kill -s SIGINT $SKYCOIN_PID
-wait $SKYCOIN_PID
+  echo "shutting down $COIN node"
 
-rm "$BINARY"
+  # Shutdown skycoin node
+  kill -s SIGINT $SKYCOIN_PID
+  wait $SKYCOIN_PID
+
+  rm "$BINARY"
+fi
 
 
 if [[ (-z $TEST || $TEST = "api") && $API_FAIL -ne 0 ]]; then
